mvp/requirements.txt
# Google Ads ChangeEvent Monitor MVP - Python Dependencies
# Minimal dependencies for MVP functionality

# Flask web framework
flask==3.0.0
flask-cors==4.0.0

# Google Ads API
google-ads==28.4.0

# Environment variables
python-dotenv==1.0.0

# Optional: For better development experience
# Uncomment if needed
# werkzeug==3.0.1  # WSGI utilities (included with Flask)
# click==8.1.7      # CLI utilities (included with Flask)
